refine cgotorch source files
