+++ cgotorch/Makefile
@@ -6,7 +6,7 @@ libtorch-macos-1.6.0.zip :
 macos/libtorch : libtorch-macos-1.6.0.zip
 	unzip -qq -o $< -d macos
 
-libcgotorch.so : cgotorch.cc cgotorch.h macos/libtorch
+libcgotorch.so : torch.cc functional.cc init.cc optim.cc cgotorch.h macos/libtorch
 	rm -f libtorch
 	ln -s macos/libtorch libtorch
 	clang++ -std=c++14 \
@@ -16,7 +16,7 @@ libcgotorch.so : cgotorch.cc cgotorch.h macos/libtorch
 	-L libtorch/lib \
 	-fPIC \
 	-shared \
-	cgotorch.cc \
+	torch.cc functional.cc init.cc optim.cc \
 	-o $@ -install_name @rpath/$@ \
 	-Wl,-rpath,libtorch/lib \
 	-Wl,-all_load libtorch/lib/libc10.dylib \
