+++ cgotorch/cgotorch.h
@@ -18,6 +18,12 @@ typedef void *Normalize;
 #endif
 
 ////////////////////////////////////////////////////////////////////////////////
+// Helper functions
+////////////////////////////////////////////////////////////////////////////////
+
+const char *exception_str(const std::exception &e);
+
+////////////////////////////////////////////////////////////////////////////////
 // Tensor construction and operations, torch
 ////////////////////////////////////////////////////////////////////////////////
