+++ cgotorch/torch.cc
@@ -1,0 +1,199 @@
+// Copyright 2020, GoTorch Authors
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <vector>
+
+#include "torch/script.h"
+#include "torch/torch.h"
+
+// FIXME(shendiaomo): including cgotorch.h before torch/torch.h will fail
+#include "cgotorch/cgotorch.h"
+
+////////////////////////////////////////////////////////////////////////////////
+// Helper functions
+////////////////////////////////////////////////////////////////////////////////
+
+const char *exception_str(const std::exception &e) {
+  auto len = strlen(e.what());
+  auto r = new char[len + 1];
+  snprintf(r, len, "%s", e.what());
+  return r;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Tensor construction and operations
+////////////////////////////////////////////////////////////////////////////////
+
+const char *RandN(int64_t *size, int64_t length, int64_t require_grad,
+                  Tensor *result) {
+  try {
+    at::Tensor t =
+        torch::randn(torch::IntArrayRef(size, length),
+                     at::TensorOptions().requires_grad(require_grad));
+    *result = new at::Tensor(t);
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *Empty(int64_t *size, int64_t length, int64_t require_grad,
+                  Tensor *result) {
+  try {
+    at::Tensor t =
+        torch::empty(torch::IntArrayRef(size, length),
+                     at::TensorOptions().requires_grad(require_grad));
+    *result = new at::Tensor(t);
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *MM(Tensor a, Tensor b, Tensor *result) {
+  try {
+    at::Tensor c = at::mm(*a, *b);
+    *result = new at::Tensor(c);
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *Sum(Tensor a, Tensor *result) {
+  try {
+    *result = new at::Tensor(a->sum());
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *Relu(Tensor a, Tensor *result) {
+  try {
+    *result = new at::Tensor(a->relu());
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *LeakyRelu(Tensor a, double negative_slope, Tensor *result) {
+  try {
+    *result = new at::Tensor(at::leaky_relu(*a, negative_slope));
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *Tanh(Tensor a, Tensor *result) {
+  try {
+    *result = new at::Tensor(a->tanh());
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *Sigmoid(Tensor a, Tensor *result) {
+  try {
+    *result = new at::Tensor(a->sigmoid());
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *View(Tensor a, Tensor *result, int64_t *size, int64_t size_len) {
+  try {
+    *result = new at::Tensor(a->view(torch::IntArrayRef(size, size_len)));
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+const char *LogSoftmax(Tensor a, int64_t dim, Tensor *result) {
+  try {
+    *result = new at::Tensor(a->log_softmax(dim));
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+void Tensor_Print(Tensor a) { std::cout << *a << std::endl; }
+
+void Tensor_Close(Tensor a) { delete a; }
+
+// The caller must free the returned string by calling FreeString.
+const char *Tensor_String(Tensor a) {
+  std::stringstream ss;
+  ss << *a;
+  std::string s = ss.str();
+  char *r = new char[s.size() + 1];
+  snprintf(r, s.size(), "%s", s.c_str());
+  return r;
+}
+
+void FreeString(const char *s) { delete[] s; }
+
+// Backward, Gradient
+void Tensor_Backward(Tensor a) { a->backward(); }
+Tensor Tensor_Grad(Tensor a) { return new at::Tensor(a->grad()); }
+
+////////////////////////////////////////////////////////////////////////////////
+//  Dataset, DataLoader, and Iterator torch.utils.data
+////////////////////////////////////////////////////////////////////////////////
+
+const char *Dataset_MNIST(const char *data_root, Dataset *dataset) {
+  try {
+    dataset->p = new torch::data::datasets::MNIST(std::string(data_root));
+    return nullptr;
+  } catch (const std::exception &e) {
+    return exception_str(e);
+  }
+}
+
+void MNIST_Close(Dataset d) { delete d.p; }
+
+void Dataset_Normalize(Dataset *dataset, double mean, double stddev) {
+  dataset->normalize = new torch::data::transforms::Normalize<>(mean, stddev);
+}
+
+using TypeMapDataset = torch::data::datasets::MapDataset<
+    torch::data::datasets::MapDataset<torch::data::datasets::MNIST,
+                                      torch::data::transforms::Normalize<>>,
+    torch::data::transforms::Stack<torch::data::Example<>>>;
+using TypeDataLoader =
+    torch::data::StatelessDataLoader<TypeMapDataset,
+                                     torch::data::samplers::SequentialSampler>;
+
+using TypeIterator = torch::data::Iterator<TypeDataLoader::BatchType>;
+DataLoader MakeDataLoader(Dataset dataset, int64_t batchsize) {
+  auto map_dataset = dataset.p->map(*dataset.normalize)
+                         .map(torch::data::transforms::Stack<>());
+  auto p = torch::data::make_data_loader(map_dataset, batchsize);
+  return p.release();
+}
+
+Iterator Loader_Begin(DataLoader loader) {
+  return new TypeIterator(static_cast<TypeDataLoader *>(loader)->begin());
+}
+
+void Iterator_Batch(Iterator iter, Tensor *data, Tensor *target) {
+  auto i = *static_cast<TypeIterator *>(iter);
+  *data = new at::Tensor(i->data);
+  *target = new at::Tensor(i->target);
+}
+
+bool Loader_Next(DataLoader loader, Iterator iter) {
+  return ++*static_cast<TypeIterator *>(iter) !=
+         static_cast<TypeDataLoader *>(loader)->end();
+}
+
+void Loader_Close(DataLoader loader) {
+  delete static_cast<TypeDataLoader *>(loader);
+}
